% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a11_catchment_summed}
\alias{a11_catchment_summed}
\title{A*11:01 Catchment Summed}
\format{
An object of class \code{list} of length 2.
}
\source{
delNero2022
}
\usage{
a11_catchment_summed
}
\description{
A*11:01 Catchment Summed
}
\seealso{
data-raw/add_catchment_calculations_and_data.R
}
\keyword{datasets}
