% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_country_census_tract_allele_frequencies.R
\name{calculate_country_census_tract_allele_frequencies}
\alias{calculate_country_census_tract_allele_frequencies}
\title{Calculate country census tract allele frequencies}
\usage{
calculate_country_census_tract_allele_frequencies(query_allele)
}
\arguments{
\item{query_allele}{The allele to query for frequencies.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{us_population_race_code_percentages_by_tract}: A data frame with the allele frequencies for each census tract.
\item \code{sf_tract_data_for_all_states}: A data frame with additional data for each census tract.
}
}
\description{
This function calculates the allele frequencies for each census tract in all states of the country.
}
\examples{
\dontrun{
calculate_country_census_tract_allele_frequencies("A")
}

}
