% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_adjusted_nmdp_hla_frequencies_by_state}
\alias{census_adjusted_nmdp_hla_frequencies_by_state}
\title{census_adjusted_nmdp_hla_frequencies_by_state}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 608660 rows and 11 columns.
}
\source{
data-raw/census_adjusted_nmdp_hla_frequencies_by_state.R and obtain_census_adjusted_nmdp_hla_frequencies_by_region
}
\usage{
census_adjusted_nmdp_hla_frequencies_by_state
}
\description{
\itemize{
\item US Census-adjusted per nmdp race group hla allelic frequencies by State
}
}
\section{Fields}{

\describe{
\item{\code{region}}{\itemize{
\item What region
}}

\item{\code{loci}}{\itemize{
\item what HLA loci
}}

\item{\code{allele}}{\itemize{
\item what HLA allele
}}

\item{\code{is_g}}{\itemize{
\item is group allele
}}

\item{\code{nmdp_race_code}}{\itemize{
\item nmdp race code
}}

\item{\code{nmdp_af}}{\itemize{
\item nmdp allelic frequiency
}}

\item{\code{nmdp_calc_gf}}{\itemize{
\item nmdp genotypic frequency
}}

\item{\code{fips}}{\itemize{
\item State fips
}}

\item{\code{census_region}}{\itemize{
\item census region - in this case the state
}}

\item{\code{us_2020_percent_pop}}{\itemize{
\item Percent of given race code population within the state
}}

\item{\code{us_2020_nmdp_gf}}{\itemize{
\item adjusted genotypic frequency based on race pop percentage in the state
}}
}}

\keyword{datasets}
