% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_state_info_by_census_tract.R
\name{get_all_state_info_by_census_tract}
\alias{get_all_state_info_by_census_tract}
\title{get_all_state_info_by_census_tract}
\usage{
get_all_state_info_by_census_tract()
}
\description{
Retrieves census tract information and maps for all states in the United States.
This function queries state codes, iterates through each state, and retrieves
the corresponding census tract information and map. If an error occurs during
the process, it is caught and logged, and
}
