% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmdp_hla_frequencies_by_race}
\alias{nmdp_hla_frequencies_by_race}
\title{NMDP HLA Frequencies and GF by Race code}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60866 rows and 7 columns.
}
\source{
data-raw/import_gragert2013_data.R
}
\usage{
nmdp_hla_frequencies_by_race
}
\description{
NMDP HLA Frequencies and GF by Race code
}
\section{Fields}{

\describe{
\item{\code{region}}{description of the region}

\item{\code{loci}}{HLA loci}

\item{\code{allele}}{HLA allele of interest}

\item{\code{is_g}}{is a group allele}

\item{\code{nmdp_race_code}}{the original catagory from nmdp}

\item{\code{nmdp_af}}{Alleleic frequency}

\item{\code{nmdo_calc_gf}}{Genotypic frequency (calculated from AF)}
}}

\keyword{datasets}
