% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmdp_racegroups}
\alias{nmdp_racegroups}
\title{NMDP Race Groups}
\format{
A tibble with 21 rows and 7 columns:
\describe{
\item{nmdp_broad_race_group}{Character. The broad race group as defined by NMDP.}
\item{nmdp_race_code}{Character. The race code as defined by NMDP.}
\item{Detailed Race/ Ethnic Description}{Character. A detailed description of the race or ethnic group.}
\item{Count}{Numeric. The count of individuals in this race group.}
\item{Typed C}{Numeric. The count of individuals typed for the C gene.}
\item{Typed DQB1}{Numeric. The count of individuals typed for the DQB1 gene.}
\item{Typed DRB3/4/5}{Numeric. The count of individuals typed for the DRB3/4/5 gene.}
}
}
\source{
National Marrow Donor Program (NMDP)
}
\usage{
nmdp_racegroups
}
\description{
This dataset contains the race groups as defined by the National Marrow Donor Program (NMDP).
}
\examples{
# Load the dataset
data(nmdp_racegroups)

# Display the first few rows of the dataset
head(nmdp_racegroups)
}
\keyword{datasets}
