% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/obtain_census_adjusted_nmdp_hla_frequencies_by_region.R
\name{obtain_census_adjusted_nmdp_hla_frequencies_by_region}
\alias{obtain_census_adjusted_nmdp_hla_frequencies_by_region}
\title{Obtain US Census-adjusted nmdp Alleleic Frequencies by Region}
\usage{
obtain_census_adjusted_nmdp_hla_frequencies_by_region(
  in_region,
  region_level = "state"
)
}
\arguments{
\item{in_region}{A string of either 'us', 'all states' (not useful), or a valid state name like 'Alaska'}

\item{region_level}{A string of either 'state' or 'county' - default is 'state'}
}
\value{
Data Frame of US Census Adjusted HLA Alleleic Frequencies
\itemize{
\item allele - the Allele in question
\item is_g - This trailing \code{g} on some of the alleles to include they are inclusive of multiple alleles ( see Maiers2007 for details)
\item nmdp_race_code - Value used by nmdp to denote races in their studies
\item nmdp_af - Allelic frequency measured by nmdp within a race
\item nmdp_calc_gf - Genotypic frequency calculated righ toff the allelic frequency using hardy-weinburg
\item region - used in downstream tables to denote regions - this is only US
\item census_region - function used in this function to call up US 2020 census data can be adjusted for states - if a state is chosen it is put here
\item total_2020_pop - total population in a given region
\item us_2020_percent_pop - percentage of the overall population with the given nmdp code in a given region. Should sum close to 1 for a specific HLA Allele
\item us_2020_nmdp_gf - The US-Census race-adjusted genotypic frequency value - used in downstream applications
}
}
\description{
Obtain US Census-adjusted nmdp Alleleic Frequencies by Region
}
\note{
\itemize{
\item relies on preprocessing performed in 'preprocessing/nmdpFrequencies.Rmd'
}
}
\examples{
obtain_census_adjusted_nmdp_hla_frequencies_by_region(in_region = 'Alaska')
}
