% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_county_info_by_state.R
\name{query_county_info_by_state}
\alias{query_county_info_by_state}
\title{Query County 2020 Census Codes by State}
\usage{
query_county_info_by_state(query_state_abbreviation)
}
\arguments{
\item{query_state_abbreviation}{}
}
\value{
A data frame of county codes by query state
}
\description{
Query County 2020 Census Codes by State
}
\examples{
tx_codes <- query_county_info_by_state(query_state_abbreviation = "TX")
}
