% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_state_codes.R
\name{query_state_codes}
\alias{query_state_codes}
\title{Query Census 2020 State Codes}
\usage{
query_state_codes()
}
\value{
a data frame of 2020 state codes
}
\description{
Query the 2020 state codes from the US Census Bureau
}
\examples{
query_state_codes()
}
