% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/us_pop_multirace_adjusted_in_btm_codes_by_region.R
\name{us_pop_multirace_in_nmdp_codes_by_region}
\alias{us_pop_multirace_in_nmdp_codes_by_region}
\title{Obtain US Population numbers by nmdp Racecodes and by Region}
\usage{
us_pop_multirace_in_nmdp_codes_by_region(
  in_region,
  census_region_level = "state"
)
}
\arguments{
\item{in_region}{A string of either 'us', 'all states' (not useful), or a valid state name like 'Alaska'}

\item{census_region_level}{A string of either 'state' or 'county' to specify if the data for region should be divided by state or by county inside the state. The default is state.}
}
\value{
Data Frame of per nmdp Race code population numbers per input region
\itemize{
\item region - used in downstream tables to denote regions - this is only US
\item nmdp_race_code - Value used by nmdp to denote races in their studies
\item total_single_race_pop - Sum of individuals in given region with given race code - who reported being just 1 race
\item total_multiple_race_pop - Sum of individuals in given region with given race code - who reported being >1 race
\item total_2020_pop - Sum of individuals (regardless of # of reported races), that included the given race on their report, by census_region
\item census_region - function used in this function to call up US 2020 census data can be adjusted for states - if a state is chosen it is put here
}
}
\description{
Obtain US Population numbers by nmdp Racecodes and by Region
}
\examples{
us_pop_multirace_in_nmdp_codes_by_region(in_region = 'Alaska')
}
