% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_county_census_tract_allele_frequencies.R
\name{calculate_county_census_tract_allele_frequencies}
\alias{calculate_county_census_tract_allele_frequencies}
\title{Calculate a nmdp Genotypic Frequency table by Census Tracks within a state, adjusted to 2020 Census Population Data}
\usage{
calculate_county_census_tract_allele_frequencies(
  state_abbreviation,
  query_allele
)
}
\arguments{
\item{state_abbreviation}{A State 2-letter postal code}

\item{query_allele}{What HLA you'd like plotted from nmdp}
}
\value{
a list containing data to be plotted as a table and a ggplot plot of the census adjusted GF
}
\description{
Calculate a nmdp Genotypic Frequency table by Census Tracks within a state, adjusted to 2020 Census Population Data
}
