% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_state_info_by_census_tract.R
\name{get_all_state_info_by_census_tract}
\alias{get_all_state_info_by_census_tract}
\title{get_all_state_info_by_census_tract}
\usage{
get_all_state_info_by_census_tract()
}
\value{
A list containing census tract information and map for each state.
}
\description{
Retrieves census tract information and map for all states.
}
