% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_tract_info_and_map_by_state.R
\name{get_census_tract_info_and_map_by_state}
\alias{get_census_tract_info_and_map_by_state}
\title{Get Census Tract Info and Map by State}
\usage{
get_census_tract_info_and_map_by_state(state_abbreviation, h3_resolution)
}
\arguments{
\item{h3_resolution}{}
}
\value{
an_object
}
\description{
Get Census Tract Info and Map by State
}
\examples{
\dontrun{tx_5 <- get_census_tract_info_and_map_by_state(state_abbreviation = 'TX',h3_resolution = 5)}

}
