% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_county_census_tract_allele_frequency_map.R
\name{make_county_census_tract_allele_frequency_map}
\alias{make_county_census_tract_allele_frequency_map}
\title{Make a plot by County of Census Tract Population corrected genotypic frequencies}
\usage{
make_county_census_tract_allele_frequency_map(
  state_abbreviation,
  county_code,
  county_name,
  query_allele
)
}
\arguments{
\item{query_allele}{}
}
\value{
an_object
}
\description{
Make a plot by County of Census Tract Population corrected genotypic frequencies
}
\examples{
 \dontrun{texus_fort_bend <- make_county_census_tract_allele_frequency_map(
state_abbreviation = 'TX',
query_allele = 'A*11:01',
county_code = '157',
county_name = 'Fort Bend'
)}
}
