% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_state_by_county_allele_frequency_map.R
\name{make_state_by_county_allele_frequency_map}
\alias{make_state_by_county_allele_frequency_map}
\title{Make a plot by State of County Population corrected genotypic frequencies}
\usage{
make_state_by_county_allele_frequency_map(state_name, query_allele)
}
\arguments{
\item{query_allele}{}
}
\value{
an_object
}
\description{
Make a plot by State of County Population corrected genotypic frequencies
}
\examples{
 \dontrun{ca_by_county <- make_state_by_county_allele_frequency_map(
state_name = 'Texas',
query_allele = 'A*11:01',
)}
}
