% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmdp_hla_frequencies_us_2020_census_adjusted}
\alias{nmdp_hla_frequencies_us_2020_census_adjusted}
\title{nmdp_hla_frequencies_us_2020_census_adjusted}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2341 rows and 6 columns.
}
\source{
data-raw/adjust_nmdp_for_us_census.R
}
\usage{
nmdp_hla_frequencies_us_2020_census_adjusted
}
\description{
nmdp allele frequencies  adjusted by race according to the US census. This table contains those race-adjusted alleleic frequencies summed up so we have a us-level summarization
}
\section{Fields}{

\describe{
\item{\code{region}}{\itemize{
\item What region was considered
}}

\item{\code{loci}}{\itemize{
\item -what loci A,B, or C
}}

\item{\code{allele}}{\itemize{
\item What allele
}}

\item{\code{af}}{\itemize{
\item US-census adjusted alleleic frequency (sum of us-census adjsuted by race afs)
}}

\item{\code{calc_gf}}{\itemize{
\item genotypic frequencies calcualted from the af
}}

\item{\code{hla_source}}{\itemize{
\item us_census_adjusted_nmdp_summed_by_race
}}
}}

\keyword{datasets}
