% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delNero2022_catchment_areas.R
\name{plot_delNero2022_catchment_areas}
\alias{plot_delNero2022_catchment_areas}
\title{Plot delNero2022 Catchment Areas}
\usage{
plot_delNero2022_catchment_areas(
  query_allele,
 
    sf_tract_centroids_for_all_states_with_catchment_with_us_population_race_code_percentages_by_tract_summed
)
}
\arguments{
\item{query_allele}{}
}
\value{
an_object
}
\description{
Plot delNero2022 Catchment Areas
}
\examples{
\dontrun{plot_delNero2022_catchment_areas('A*11:01')}
}
