% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/summarize_tract_genotypic_frequencies_by_h3_hexagon.R
\name{summarize_tract_genotypic_frequencies_by_h3_hexagon}
\alias{summarize_tract_genotypic_frequencies_by_h3_hexagon}
\title{Summarize Genotypic frequency by H3 Hexagon and State}
\usage{
summarize_tract_genotypic_frequencies_by_h3_hexagon(
  state_abbreviation,
  query_allele,
  h3_resolution
)
}
\arguments{
\item{state_abbreviation}{The state abbreviation, e.g. 'MA'}

\item{query_allele}{The allele to query, e.g. 'A*11:01'}

\item{h3_resolution}{The resolution of the H3 hexagon, e.g. 5}
}
\value{
an_object
}
\description{
Summarize Genotypic frequency by H3 Hexagon and State
}
\examples{
\dontrun{summarize_tract_genotypic_frequencies_by_h3_hexagon('MA','A*11:01',5)}
}
