% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_counties_geo_data.R
\name{us_counties_geo_data}
\alias{us_counties_geo_data}
\title{US Counties complementary geo data}
\usage{
us_counties_geo_data()
}
\value{
Data frame with us counties complementary
}
\description{
A function that returns county fips and GEOID information to complement census data
}
